in quality_and_coverage script, added check for ./SEQ//fastq_files DIR before copying reads from Basespace to VM. Fixed quality_and_coverage script to call run_type_pipe_2.3.sh. Changed check function to use -e instead of find in type_pipe_2.3

+++ quality_and_coverage_1.1.sh
@@ -62,7 +62,14 @@ for line in ${spp_variables[@]}; do
 
     #Copy all the files from basespace into the newly created folder.
     yes n | basemount Basespace/
-    cp ./Basespace/Projects/SEQ${SEQ_NUM}_QC_${extension}/Samples/*/Files/* SEQ${SEQ_NUM}/${species}
+    if [ -e ./SEQ${SEQ_NUM}/${species}/fastq_files ]; then
+        echo "Directory SEQ${SEQ_NUM}/${species}/fastq_files already exists, skipping transfer of reads from Basespace to local DIR"
+    else
+        echo "Copying reads from Basespace now..."
+        cp ./Basespace/Projects/SEQ${SEQ_NUM}_QC_${extension}/Samples/*/Files/* SEQ${SEQ_NUM}/${species}
+    fi
+
+    #cp ./Basespace/Projects/SEQ${SEQ_NUM}_QC_${extension}/Samples/*/Files/* SEQ${SEQ_NUM}/${species}
 
     #Check that there are actually fastq files, and if there aren't, end script and remove empty folders
     if [[ ! -z `find SEQ${SEQ_NUM}/${species} -maxdepth 1 -name "*fastq*"` ]]; then
@@ -74,7 +81,7 @@ for line in ${spp_variables[@]}; do
     fi
 
     cd SEQ${SEQ_NUM}/${species}
-    /home/staphb/scripts/run_type_pipe_2.2.sh -l "$genome_length"
+    /home/staphb/scripts/run_type_pipe_2.3.sh -l "$genome_length"
     cd
 done
